Correcciones segundo parcial: JSON para preguntas, limpieza de archivos, changelog y plantilla de PR
