+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,43 @@
+# Pull Request - [Nombre de la entrega o feature]
+
+## Descripción
+
+Descipcion breve de los cambios realizados en esta PR.
+
+## Problema que Resuelve / Observaciones Abordadas
+
+¿Qué problema o tarea aborda esta PR?
+
+## Tipo de Cambio
+
+Por favor, marca con una `x` las opciones que apliquen:
+
+- [ ] Bug fix (corrección de un error)
+- [ ] Nueva característica (adición de una nueva funcionalidad)
+- [ ] Mejora (optimización o refactorización de código existente)
+- [ ] Documentación (cambios en la documentación)
+- [ ] Configuración (cambios en archivos de configuración, scripts, etc.)
+- [ ] **Corrección de Evaluación (cambios realizados para abordar feedback de una evaluación)**
+
+## ¿Cómo se Probó?
+
+Pasos para probar los cambios.
+
+## Checklist
+
+- [ ] Mi código sigue las guías de estilo del proyecto.
+- [ ] He realizado una auto-revisión de mi propio código.
+- [ ] He comentado mi código, especialmente en áreas difíciles de entender.
+- [ ] He actualizado la documentación relevante (ej. README, CHANGELOG).
+- [ ] Mis cambios no introducen nuevas advertencias.
+- [ ] He añadido pruebas que demuestran que mi corrección o nueva característica funciona (si aplica).
+- [ ] Los cambios en el `CHANGELOG.md` están actualizados y reflejan esta PR.
+- [ ] He verificado que la aplicación se ejecuta correctamente con estos cambios.
+
+## Enlaces Relevantes
+
+Cualquier enlace a recursos externos o PRs relacionadas.
+
+## Observaciones Adicionales para el Revisor
+
+Cualquier información extra que el revisor deba saber.
