+++ api/preguntas.json
@@ -1,0 +1,67 @@
+[
+  {
+    "id": 1,
+    "texto": "¿Cómo se ve tu piel al final del día?",
+    "opciones": [
+      { "texto": "Brillante en toda la cara", "valor": "grasa" },
+      { "texto": "Brillante en la zona T, seca en mejillas", "valor": "mixta" },
+      { "texto": "Opaca y seca", "valor": "seca" },
+      {
+        "texto": "Equilibrada, sin brillo excesivo ni resequedad",
+        "valor": "normal"
+      }
+    ]
+  },
+  {
+    "id": 2,
+    "texto": "¿Cómo se siente tu piel al tacto después de lavarla?",
+    "opciones": [
+      { "texto": "Tirante y áspera", "valor": "seca" },
+      { "texto": "Suave y cómoda", "valor": "normal" },
+      { "texto": "Grasa o pegajosa", "valor": "grasa" },
+      { "texto": "Tirante en algunas áreas y grasa en otras", "valor": "mixta" }
+    ]
+  },
+  {
+    "id": 3,
+    "texto": "¿Cómo describirías tus poros?",
+    "opciones": [
+      { "texto": "Pequeños y poco visibles", "valor": "normal" },
+      { "texto": "Grandes y visibles en toda la cara", "valor": "grasa" },
+      {
+        "texto": "Grandes en la zona T, pequeños en mejillas",
+        "valor": "mixta"
+      },
+      {
+        "texto": "Pequeños pero con tendencia a enrojecerse o irritarse",
+        "valor": "sensible"
+      }
+    ]
+  },
+  {
+    "id": 4,
+    "texto": "¿Cómo reacciona tu piel a nuevos productos o cambios climáticos?",
+    "opciones": [
+      { "texto": "Se irrita fácilmente", "valor": "sensible" },
+      { "texto": "No presenta cambios notables", "valor": "normal" },
+      {
+        "texto": "Se reseca o se engrasa dependiendo del clima",
+        "valor": "mixta"
+      },
+      {
+        "texto": "Se pone más brillante con el calor y más seca con el frío",
+        "valor": "grasa"
+      }
+    ]
+  },
+  {
+    "id": 5,
+    "texto": "¿Con qué frecuencia sientes la necesidad de hidratar tu piel?",
+    "opciones": [
+      { "texto": "Varias veces al día", "valor": "seca" },
+      { "texto": "Una vez al día es suficiente", "valor": "normal" },
+      { "texto": "Rara vez necesito hidratarla", "valor": "grasa" },
+      { "texto": "Depende de la zona del rostro", "valor": "mixta" }
+    ]
+  }
+]
