+++ assets/js/pregunta.js
@@ -1,0 +1,7 @@
+export class Pregunta {
+    constructor(id, texto, opciones) {
+        this.id = id;
+        this.texto = texto;
+        this.opciones = opciones;
+    }
+}
